+++ image/build-image.sh
@@ -158,6 +158,8 @@ cp -v "${SCRIPT_DIR}/rootfs/home/pi/printer_data/config/crowsnest.conf" \
     "${ROOT_MNT}/home/pi/printer_data/config/crowsnest.conf"
 cp -v "${SCRIPT_DIR}/rootfs/home/pi/printer_data/config/crowsnest-usb.conf" \
     "${ROOT_MNT}/home/pi/printer_data/config/crowsnest-usb.conf"
+cp -v "${SCRIPT_DIR}/rootfs/home/pi/printer_data/config/nfc_spoolman.cfg" \
+    "${ROOT_MNT}/home/pi/printer_data/config/nfc_spoolman.cfg"
 
 # Create gcode directory
 mkdir -p "${ROOT_MNT}/home/pi/printer_data/gcodes"